requirements.txt
# Azure SDK - Document Intelligence
azure-ai-documentintelligence>=1.0.0b1

# Azure SDK - OpenAI
openai>=1.12.0
azure-identity>=1.15.0

# Data validation and modeling
pydantic>=2.5.0
pydantic-settings>=2.1.0

# UI Framework
streamlit>=1.30.0

# Utilities
python-dotenv>=1.0.0
pillow>=10.0.0

# Logging
structlog>=24.0.0
